Traces list in side-panel
- Split file SummaryPanelNode => (SummaryPanelNode + SummaryPanelNodeMetrics)
- When traces are enabled & service is selected, show a tabbed view Traces/Metrics below contextual node info

Fixes https://github.com/kiali/kiali/issues/2840

Load traces list in side panel

- Share some formatting stuff with TracesDetails
- Load traces & create TraceListItem component
- Rename Metrics => Traffic for newly created tab

Use side-panel standard font size

Tie traces list with graph refresh

Button font size

URL-based experimental flags

Experimental feature switch based on "xflags" URL param
For In Graph Tracing, add to URL: "&xflags=igt"

Style: ellipsis & bottom margin
