+++ src/components/JaegerIntegration/JaegerResults/JaegerItem.tsx
@@ -6,12 +6,12 @@ import { JaegerTrace, Span } from '../../../types/JaegerInfo';
 import { JaegerTraceTitle } from './JaegerTraceTitle';
 import { SpanDetail } from './SpanDetail';
 import { style } from 'typestyle';
-import moment from 'moment';
-import { formatRelativeDate, cleanServiceSelector } from './transform';
+import { cleanServiceSelector } from './transform';
 import { isErrorTag } from '../RouteHelper';
 import { KialiAppState } from '../../../store/Store';
 import { connect } from 'react-redux';
 import { PfColors } from '../../Pf/PfColors';
+import { getFormattedTraceInfo } from './FormattedTraceInfo';
 
 const labelStyle = style({
   margin: '5px'
@@ -56,34 +56,25 @@ class JaegerItemC extends React.Component<JaegerScatterProps, JaegerScatterState
   };
 
   render() {
-    const { trace } = this.props;
-    const { duration, services, spans, startTime } = trace;
-    const numSpans = spans.length;
-    const mDate = moment(startTime / 1000);
-    const fromNow = mDate.fromNow();
-    const timeStr = mDate.format('h:mm:ss a');
-    const numErredSpans = spans.filter(sp => sp.tags.some(isErrorTag)).length;
+    const { trace, jaegerURL } = this.props;
+    const formattedTrace = getFormattedTraceInfo(trace);
     return (
       <Card isCompact style={{ border: '1px solid #e6e6e6' }}>
         <JaegerTraceTitle
-          trace={trace}
-          duration={duration}
-          onClickLink={this.props.jaegerURL !== '' ? `${this.props.jaegerURL}/trace/${this.props.trace.traceID}` : ''}
+          traceID={trace.traceID}
+          formattedTrace={formattedTrace}
+          onClickLink={jaegerURL !== '' ? `${jaegerURL}/trace/${trace.traceID}` : ''}
         />
         <CardBody>
           <Grid style={{ marginTop: '20px' }}>
             <GridItem span={2}>
-              <Label>
-                {numSpans} Span{numSpans > 1 && 's'}
-              </Label>
-              {Boolean(numErredSpans) && (
-                <Label style={{ marginLeft: '10px', backgroundColor: PfColors.Red200 }}>
-                  {numErredSpans} Error{numErredSpans > 1 && 's'}
-                </Label>
+              <Label>{formattedTrace.spans}</Label>
+              {formattedTrace.errors && (
+                <Label style={{ marginLeft: '10px', backgroundColor: PfColors.Red200 }}>{formattedTrace.errors}</Label>
               )}
             </GridItem>
             <GridItem span={8}>
-              {sortBy(services, s => s.name).map(service => {
+              {sortBy(trace.services, s => s.name).map(service => {
                 const { name, numberOfSpans: count } = service;
                 const spans = trace.spans.filter(span => span.process.serviceName === name);
                 const errorSpans = spans.filter(span => span.tags.some(isErrorTag)).length;
@@ -100,11 +91,11 @@ class JaegerItemC extends React.Component<JaegerScatterProps, JaegerScatterState
               })}
             </GridItem>
             <GridItem span={2} style={{ textAlign: 'right' }}>
-              {formatRelativeDate(startTime / 1000)}
+              {formattedTrace.relativeDate}
               <span style={{ padding: '0 10px 0 10px' }}>|</span>
-              {timeStr.slice(0, -3)}&nbsp;{timeStr.slice(-2)}
+              {formattedTrace.absTime}
               <br />
-              <small>{fromNow}</small>
+              <small>{formattedTrace.fromNow}</small>
             </GridItem>
             {this.state.spansSelected.length > 0 && (
               <GridItem span={12}>
